issue #11 - add unit tests * refactored a little bit
